Moved HardMode inside GamePage folder and added a declarations file inside src for .mp3 imports.
