+++ Home-Page/src/declarations.d.ts
@@ -1,0 +1,4 @@
+declare module "*.mp3" {
+    const src: string;
+    export default src;
+  }
